BUGFIX: Timer switch solution used wall clock, now switched to process time.
This fix coincides with the solution used in the master branch.

+++ src/switches.c
@@ -853,7 +853,7 @@ switcher (const int process, int index, int commandline)
 	}
       else
 	{
-	  set_time_limit ((double) integer_argument ());
+	  set_time_limit (integer_argument ());
 	  return index;
 	}
     }
